Makefile: Add initial build environment files
- Typical .gitignore
- GNU Makefile
- Crosstool-NG recursion patch (old, but still serves its purpose)

Crosstool-NG is a fundational piece to all of this. Its narrow focus on
generating toolchains is exactly what is wanted in this context.

http://crosstool-ng.github.io/

Also, the Makefile explicitly checks for Ubuntu Bionic Beaver 18.04 LTS.
The host OS will only change when Ubuntu releases its next LTS. This can
be tinkered with of course.