etinker: Refactor GNU make function usage with high-level targets
Leverage function's ability to accept arguments to avoid abuse of the
make rule parameters. There was some shuffling of targets in Makefile to
handle the fact that '.config' files and their original source can be
touched but not actually changed. This would trigger rebuild of the
toolchain and kernel without need. This change should provide a pattern
that is consistent for other etinker components that will be added in
the near future. Especially those that use 'Kconfig' style
configuration infrastructure.