README: add AML-S905X-CC (Le Potato)
This board boots upstream U-Boot and Linux with etinker. The
'bootloader-sync-mmc' target will write the appropriate U-Boot binary
to the mounted SD/MMC disk. A '/usr/sbin/aml-update' script is provided,
on the embedded target, to update U-Boot on the eMMC also.

This board relys on binary blobs, and proprietary tools, to generate
working U-Boot images. Access to SOC reference manuals was missing. This
was an exercise in persistence. I am not sure that I really learned much
about the actual SOC.

The biggest learning area was dealing with the ARM Cortex-A53 (AArch64)
architecture. The 64-bit toolchain builds a working embedded Linux stack.

It boots!