etinker: refactor 'defconfig' handling
Hopefully, this commit will simplify how defconfig files are used going
forward.

[Linux]
When a board is added, the 'et_<board_type>_defconfig' should be created
per the needs of the current board. If a subsequent board of the same
type is added later, then 'make savedefconfig' will update the common
'defconfig' for all boards.

[U-Boot]
When a board is added, the 'et_<board>_defconfig' should be created per
the needs of the current board. Each board's 'defconfig' is unique for
U-Boot.