scripts/mksdcard: drop 'tiny' SD/MMC formatting
After some experiments with random 1, 2, and 4 GB cards, it became very
apparent that making these work was not worth my time anymore. It is/was
a nice thought. The variances in SD Class 2, 4, and 6 devices is more
than I want to manage anymore. It is more worthwhile to fix things on
modern media. I will just need to spend a few more $$$.

NOTE:
This is not about partitioning or formatting tools. These SD/MMC cards
are just older and slower. With some quality problems along the way.