boards/ls1012afrdm: add NXP Layerscape FRDM-LS1012A board support
This no longer supported board is being added to experiment with one of
the earliest Freescale/NXP Layerscape ARM SOC implementations. While it
has a single-core 64-bit Arm(R) Cortex(R)-A53, like later chips, this
is a very different beast.

The Layerscape platform is for networking equipment. This is not a
platform for general Arm(R) application processor workloads. Spending
some time with the development tools and general platform guides will
show that this is not a hobbyist Arm(R) SOC.