buildroot: busybox: configure applets to be more desktop-like
This is the first step towards a reduction in "rootfs.tar" sizes.
The etinker images are larger than I need. I really do not use most
of the content. Most focus is typically on trying newer U-Boot and
Linux releases. This commit will allow me to test applets before
dropping the Buildroot packages.