kernel, packages: append INSTALL_MOD_PATH with '/usr'
The transition to using a symbolic link, for 'lib', in the root
directory, requires that module installation paths also get updated.
This was noticed during the transfer of files to the target with rsync.
A full '/lib/modules/x.y.z' path was being created. This leads to all
binaries failing due to the loss of the symbolic link to '/usr/lib'.