u-boot: omap3-*: use board specific FIT addresses
Place the FDT blob at the last 1 MiB of RAM. For these 256 MiB omap3-*
boards, that is 0x8ff00000. The FIT image is loaded to 0x8b000000 for
beagle-xm and evm.

omap3-beagle has 128 MiB RAM variants that are no longer supported for
booting Linux by U-Boot. I happen to have such a board. I attempt to
boot the FIT image and if it fails, I set environment variables that are
more friendly to the 128 MiB variant. If everything else is OK, it
will boot. The FIT image is loaded to 0x86000000 to cover both 128 and
256 MiB variants that I have.

I will likely doing more with FIT images going forward. Mostly to learn
how it works. It may become required for further maintenance activity.