omap3: use '-base' kernel variant for older OMAP3xxx boards
The AM335x is a bit better maintained and some differences that have
allowed it diverge in some interesting ways. This will help isolate
some of those diffences. There is more tinkering to do here.

The '-base' is intended to cover OMAP35xx to DM37xx. (I have older
Beagle Boards on hand)